+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for marking positional message arguments as optional when using annotation parsing.
 
+### Changed
+- [tanjun.annotations.Flag][] will no longer unset any previously set aliases.
+
 ### Fixed
 - Don't ignore parameter annotations when the top-level type-hint is a union.